#3 [refactor] 스토리텔링 content 공통 ui 위치 변경
