+++ feature/onboarding/src/main/java/com/sopetit/onboarding/storytelling/StoryTellingSecondScreen.kt
@@ -19,7 +19,7 @@ import androidx.compose.ui.unit.dp
 import com.sopetit.design_system.Gray50
 import com.sopetit.design_system.R
 import com.sopetit.design_system.StoryTellingContent2
-import com.sopetit.ui.common.StoryTellingText
+import com.sopetit.ui.common.item.StoryTellingText
 
 @Composable
 fun StoryTellingSecondScreen(
