+++ feature/onboarding/src/main/java/com/sopetit/onboarding/storytelling/StoryTellingThirdScreen.kt
@@ -17,7 +17,7 @@ import androidx.compose.ui.unit.dp
 import com.sopetit.design_system.Gray50
 import com.sopetit.design_system.R
 import com.sopetit.design_system.StoryTellingContent3
-import com.sopetit.ui.common.StoryTellingText
+import com.sopetit.ui.common.item.StoryTellingText
 
 @Composable
 fun StoryTellingThirdScreen(
